requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-postgres-copy>=2.4,<2.5
djangorestframework>=3.11,<3.12
pandas>=1.0.3,<1.0.5
